🚚 [variables] Move module `binders` to `domain`
